% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obstetrics-package.R
\docType{package}
\name{obstetrics-package}
\alias{obstetrics}
\alias{obstetrics-package}
\title{obstetrics: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
